unified_doc_tool/requirements.txt
Flask==3.1.2
PyMuPDF==1.27.1
pillow==12.1.1
Werkzeug==3.1.5
